refactor: api improvements
