+++ src/main/kotlin/com/je_martinez/demo/features/todos/services/TodoService.kt
@@ -1,12 +1,12 @@
 package com.je_martinez.demo.features.todos.services
 
-import com.je_martinez.demo.features.todos.cache.TodosCacheSettings
 import com.je_martinez.demo.database.models.Todo
 import com.je_martinez.demo.database.repository.TodoRepository
+import com.je_martinez.demo.features.todos.cache.TodosCacheSettings
+import com.je_martinez.demo.features.todos.dtos.requests.TodoRequest
+import com.je_martinez.demo.features.todos.dtos.responses.TodoResponse
+import com.je_martinez.demo.features.todos.dtos.responses.toResponse
 import com.je_martinez.demo.features.todos.exceptions.TodoExceptions
-import com.je_martinez.demo.features.todos.dtos.TodoRequest
-import com.je_martinez.demo.features.todos.dtos.TodoResponse
-import com.je_martinez.demo.features.todos.dtos.toResponse
 import org.bson.types.ObjectId
 import org.springframework.cache.annotation.CacheEvict
 import org.springframework.cache.annotation.CachePut
